+++ CHANGES.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Fix previews incorrectly using disposed textures after resetting settings
+* Fix wrong team exported as winning team
 
 ### Removed